Modified Frame rate for Matplotlib Redraw
Graph components now redraw at a separate rate from simulator ticks,
so we can update them less frequently.  This gives time for rendering,
for example the spectrogram.  If the UI seems choppy, decrease the ratio
for redrawing the graph components.  I also removed the xy graph that
was attached to the spectrogram, and cleared the spectrogram between
drawings, and on reset.

Fixes Issue #42